# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    name: IT Support
  description: IT Broadcast Operations & Helpdesk System API. Provides endpoints for
    ticket management, knowledge base (Big Book), shift scheduling, and real-time
    notifications.
  license:
    name: MIT
  title: G-Report API
  version: "1.0"
paths:
  /auth/login:
    get:
      description: Display the login page
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      summary: Show login form
      tags:
      - Auth
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate user with email/username and password
      parameters:
      - description: Email or Username
        in: formData
        name: email
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: Redirect to dashboard
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
      summary: User login
      tags:
      - Auth
...
  /staff:
    get:
      description: Display staff dashboard with active tickets and routines
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Staff dashboard
      tags:
      - Staff
  /staff/bigbook:
    get:
      description: Display Big Book articles and candidates
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Knowledge base page
      tags:
      - Staff
  /staff/bigbook/search:
    get:
      description: Search Big Book articles and return JSON results
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            items:
              type: object
            type: array
      security:
      - CookieAuth: []
      summary: Search knowledge base JSON
      tags:
      - Staff
  /staff/history:
    get:
      description: Display resolved tickets history for current staff
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Staff work history
      tags:
      - Staff
  /staff/tickets/{id}:
    get:
      description: Display ticket details with chat history
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
        "404":
          description: Ticket not found
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Ticket detail
      tags:
      - Staff
  /staff/tickets/{id}/resolve:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Mark ticket as resolved with solution
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: Solution description
        in: formData
        name: solution
        required: true
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: Redirect to dashboard
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Resolve ticket
      tags:
      - Staff
  /staff/tickets/list:
    get:
      description: HTMX partial for active tickets list
      produces:
      - text/html
      responses:
        "200":
          description: HTML partial
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Active tickets list
      tags:
      - Staff
securityDefinitions:
  CookieAuth:
    in: cookie
    name: user_id
    type: apiKey
swagger: "2.0"